requirements.txt
streamlit
langchain
langchain-community
python-dotenv
pdfplumber
requests
faiss-cpu